requirements.txt
openai>=0.27.8
PyYAML>=6.0.1
python-decouple>=3.8